[refactor]: Ajustando estrutura de pastas ao adicionar pasta de elementos graficos
